Struct scinotation::SciValue

pub struct SciValue<BASEVAL, EXPSTORE: SignedInt> {
    // some fields omitted
}

Structure storing a number in a format similar to scientific notation.

Probably inefficient, but also probably quite accurate.

Methods

impl<B: Int, E: SignedInt> SciValue<B, E>

fn wrap(val: B) -> SciValue<B, E>

fn wrap_with_exponent(val: B, exp: E) -> SciValue<B, E>

impl<B: Int, E: SignedInt> SciValue<B, E>

fn pow(self, exp: E) -> SciValue<B, E>

impl<B: Int + FromPrimitive, E: SignedInt> SciValue<B, E>

fn to_full_value(self) -> Option<B>

Converts the number to its full representation, out of scientific notation. Returns None if an exact representation is not possible.

An exact representation is not possible when the exponent is < 0, and when the full representation would be greater than the maximum value of the base type.

fn reduce(&self) -> SciValue<B, E>